wlan: Filter channels based on gRoamIntraBand in driver
Currently if ini parameter gRoamIntraBand is enabled in driver
sta should roam within the band 5G to 5G or 2.4G to 2.4G only,
but sta is able to roam from 5G to 2.4G and vice versa because
host sends channels of all bands to fw.

Fix is to filter channels based on gRoamIntraBand in driver
before sending channel list to fw.

Change-Id: I54407954dbf44fdc1779ce864767318e24db5757
CRs-Fixed: 2043309
Signed-off-by: FireLord1 <firelord.xda@lineageos.org>